working on creqs->ini->debugger. refactored debugger - * debugger: integrated driver is now isolated in subshell with callbacks to the script * debugger: paint now uses win* functions to draw in a clipping window * bg_ini.sh: with bgawk, scripts can not return the value 1 so changed all scripts to return 2 to indicate negative results * bg_cui.sh: added csi* string functions which ignore CSI sequences in calculating strlen. * bg_cui.sh: added win* functions which implement a buffered clipping window * bg_strings.sh: refactored stringSplit and wrote unit tests for it. * bgOptionGetOpt: bug fix: passing through -o "" would result in -o"" which is invalid. now uses -o-- * refined command_not_found_handle() to not throw assertError which was unable to end the script with bgExit --complete for some reason not yet undertood * unitTests: added normWhitespace directive * added bg_strings.sh.ut * added bg_cui.sh.ut *

+++ lib/bg_bgawk.sh
@@ -211,7 +211,7 @@ function bgawk()
 			[ "$inplace" ] && assertError -e "$exitCode" -v file -v tmpFile -v scriptFile -v exitCode -f output "the awk script returned a non-zero exit code in inplace edit mode (-i). Exit code 1 may mean that the script had a syntax error "
 			[ ! "$quietFlag" ] && [ $exitCode -eq 1 ] && assertError -e "$exitCode"  -v scriptFile -v exitCode "likely error in awk script"
 			rm "$scriptFile"
-			rm "$tmpFile"
+			[ "$tmpFile" ] && rm "$tmpFile"
 			return $exitCode
 		fi
